Refactor featureType checks to use switch expression
The previous use of individual if statement for each featureType was not concise and efficient. It has been refactored to leverage Java's switch expression which makes the code more read-able, clean, easy to maintain and manage. This change does not affect the functionality of the system but enhances the code quality.

Signed-off-by: Kinsley Kajiva <kajivakinsley@gmail.com>