Add utility method for FeatureType enum conversion
Modified the FeatureTypes by adding a new utility method `getEnumByString` to get the FeatureType enum value by a string.

This method is utilized in the ClientWebSocket class to remove the hard-coded string values used to set the FeatureType. Now, the featureType string retrieves the FeatureTypes enum value, and this enum value is used in the switch-case to set the feature type.

This increases code readability and maintainability, by preventing potential errors due to typos in hardcoded strings.

Signed-off-by: Kinsley Kajiva <kajivakinsley@gmail.com>