Improve code readability in ClientWebSocket.java
Reformatted the code in ClientWebSocket.java to abide by standard coding practices. Improvements include better alignment of variable assignments and function calls, better spacing between code blocks, and reformatting of a switch case for better legibility which is important to make the code easier to read and maintain for further updates. The logic of the code stays the same.

Signed-off-by: Kinsley Kajiva <kajivakinsley@gmail.com>