Refactor VideoCallSocketEvent to use Request constants
This change introduces constants for the request type strings in the VideoCallSocketEvent class. This is to ensure consistency and reduce the chance of bugs due to typos. Magic numbers were removed and inline hardcoded strings were replaced by defined constants. This modification helps in creating a central source of truth for all request types.

Signed-off-by: Kinsley Kajiva <kajivakinsley@gmail.com>