Update app configuration and enhance error handling
This commit updates the application configuration and improves error handling for the case when config files are not found. It changes the app configuration folder name and ignores this folder in git. It also includes a new method to ensure configuration files are copied correctly and handles cases when files are not found by expanding default configuration templates.

Signed-off-by: kinsleykajiva <kajivakinsley@gmail.com>